Add files via upload
An early version of the Python file that takes in the original dataset and cleans it, outputting a cleaned dataset. The imported csv library that I used in this works for small files, but big files will cause memory problems when you try to run this. If we can't find a machine that can handle using the Python csv library, I'll have to later find a way to clean the data without using the csv library, since it uses such wasteful csv functions. This is hard, however, because the original dataset has some values that have commas in them, making it horrendously difficult to simply just use Python to split each line by commas.