Preliminary draft of lift correlation
This code prints out the most correlated object (highest lift) from one attribute and another. It will print out the each element in the first attribute, the element in the second attribute with the highest lift value, and the lift value itself. May have to do a rework on this to see what companies are negatively correlated. May also have to clean this code up later.